Eliminates debug, again closes #85
This solves de problem as far as it can go. An alternative would be to make it work in private repos all the way, but it would need some additional checking and additional configuration for the user. This is a judgment call, that makes at least a part of the functionality work as is, and it's minimally viable